Switch Costas scan to max-log 8-FSK metric; float --min-score (default 3.0)
Replace the nearest-neighbor frequency/temporal diff scoring with the
statistically correct max-log 8-FSK metric: per symbol, expected Costas
tone minus best competitor across all 7 other tones.  This eliminates
edge-tone bias (sm=0 penalty), drops the statistically unsound temporal
comparison, and aligns the Costas score units with the soft-symbol LLRs
used by the LDPC decoder.

Also change --min-score to accept a float (std::stof, propagated through
FT8Cuda and ft8_gpu_scan to the kernel), so thresholds like 2.5 are
meaningful since the score is now a float division.  Default updated to
3.0 based on observed detection improvement.

+++ gm/cuda/FT8Cuda.cc
@@ -27,7 +27,7 @@
 namespace gm {
 namespace cuda {
 
-FT8Cuda::FT8Cuda(gm::buffer::BufferPosition<std::complex<float>>* inP, bool corpus, int min_score) :
+FT8Cuda::FT8Cuda(gm::buffer::BufferPosition<std::complex<float>>* inP, bool corpus, float min_score) :
 inPos(inP),
 buff_pos{0},
 ring_write_idx(0),
